POC TemplateRenderMixin
